+++ modules/sdk-coin-ton/test/unit/tokenTransferBuilder.ts
@@ -9,6 +9,7 @@ import * as utils from '../../src/lib/utils';
 describe('Ton Token Transfer Builder', () => {
   const tokenName = 'tton:ukwny-us';
   const factory = new TransactionBuilderFactory(coins.get(tokenName));
+
   it('should build a unsigned transfer tx', async function () {
     const txId = testData.signedTokenSendTransaction.txId.replace(/\//g, '_').replace(/\+/g, '-');
     const txBuilder = factory.getTokenTransferBuilder();
